requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
catboost>=1.2.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
statsmodels>=0.14.0

# Model Interpretation
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
